Adds base file.
This commit includes the following things:

[Global]
Extension string.Center: Center text within a specified length of space.
Extension string.Truncate: Truncates a string to the desired length,

[Console]
class ChangingOutput: This creates a neat little status box to the left
of your text and updates while an operation is in progress.